src/main/java/com/example/w23comp1008s2w10/Rectangle.java
package com.example.w23comp1008s2w10;

public class Rectangle {
    private double height, width;

    public Rectangle(double height, double width) {
        this.height = height;
        this.width = width;
    }

    public double getHeight() {
        return height;
    }

    public void setHeight(double height) {
        this.height = height;
    }

    public double getWidth() {
        return width;
    }

    public void setWidth(double width) {
        this.width = width;
    }

    public double getArea()
    {
        //rectangle is 20 x 30, so area should be 600
        return width * height;
    }
}
